bug(validation): fix bugs in validation logic - fix bug in recipe id validation - fix bug in user id validation - refactor test suites [Fixes #154526867]
