+++ client/src/components/Favorites/View.jsx
